requirements.txt
streamlit==1.44.0
praw==7.8.1
google-genai==1.8.0
requests==2.32.3
pandas==2.2.3
numpy==2.2.4
